Fix coverity warnings from static analysis tools
1.DO_NOT_USE_INT_IN_FOR_LOOP
   a.NativeSecureElement.cpp: NxpNfc_ParsePlatformID()

2.INVALID_ASSIGNMENT_WITH_UNSIGNED_VALUES
   a.NativeNfcTag.cpp: nativeNfcTag_doTransceive()

+++ nci/SN100x/jni/NativeSecureElement.cpp
@@ -485,7 +485,7 @@ static void NxpNfc_ParsePlatformID(const uint8_t* data) {
   uint8_t PlatfStrings[][MAX_STRING_SIZE] = {
       "NFCC HW is SN100", "NFCC HW is SN110", "Not SN1xx NFCC"};
 
-  uint32_t count = 0;
+  size_t count = 0;
 
 #define MAX_PLATFTYPE_ELEMENTS (sizeof(PlatfType) / sizeof(PlatfType[0]))
 #define MAX_STRINGS (sizeof(PlatfStrings) / sizeof(PlatfStrings[0]))
